move CustomModel classes to own file
